moving non-zip-related methods to utils
